Clarify form runtime option effects
Replace the vague form effects components with FormRuntimeEffects, which describes that this provider coordinates runtime side effects around the rendered form.

Keep the options feature responsible for selecting option-capable derived nodes and split the per-node option effects into RunOptionsEffectsForNode. This avoids a generic layout derivation helper growing feature-specific option behavior and makes the runner/per-node responsibilities explicit.

Delete the old FormEffects component instead of leaving aliases, so the API surface reflects the intended structure directly.
