+++ src/App/frontend/src/features/options/RunOptionsEffects.tsx
@@ -1,104 +1,35 @@
 import React from 'react';
 
 import { FormStore } from 'src/features/form/FormContext';
-import { EffectPreselectedOptionIndex } from 'src/features/options/effects/EffectPreselectedOptionIndex';
-import { EffectRemoveStaleValues } from 'src/features/options/effects/EffectRemoveStaleValues';
-import { EffectSetDownstreamParameters } from 'src/features/options/effects/EffectSetDownstreamParameters';
-import { EffectStoreLabel } from 'src/features/options/effects/EffectStoreLabel';
-import { EffectStoreLabelInGroup } from 'src/features/options/effects/EffectStoreLabelInGroup';
-import { useFetchOptions, useFilteredAndSortedOptions } from 'src/features/options/useGetOptions';
-import { useIsHidden } from 'src/utils/layout/hidden';
-import type { OptionsValueType } from 'src/features/options/useGetOptions';
-import type { IDataModelBindingsForGroupCheckbox } from 'src/layout/Checkboxes/config.generated';
-import type { IDataModelBindingsOptionsSimple } from 'src/layout/common.generated';
-import type { CompIntermediate, CompWithBehavior } from 'src/layout/layout';
-import type { IDataModelBindingsForGroupMultiselect } from 'src/layout/MultipleSelect/config.generated';
-import type { DerivedLayoutNode } from 'src/utils/layout/deriveLayoutNodes';
+import { RunOptionsEffectsForNode } from 'src/features/options/RunOptionsEffectsForNode';
+import { getComponentBehaviors, getComponentDef } from 'src/layout';
+import { DataModelLocationProviderFromRowContexts } from 'src/utils/layout/DataModelLocation';
+import { deriveLayoutNodes } from 'src/utils/layout/deriveLayoutNodes';
 
-interface RunOptionEffectsProps {
-  valueType: OptionsValueType;
-  node: DerivedLayoutNode;
-}
+export function RunOptionsEffects() {
+  const nodes = FormStore.raw.useMemoSelector((state) =>
+    deriveLayoutNodes(state).filter((node) => {
+      if (!getComponentBehaviors(node.intermediateItem.type)?.canHaveOptions) {
+        return false;
+      }
 
-export function RunOptionsEffects({ valueType, node }: RunOptionEffectsProps) {
-  const isHidden = useIsHidden(node.baseId, { respectPageOrder: true });
-
-  if (isHidden) {
-    return null;
-  }
-
-  return (
-    <RunVisibleOptionsEffects
-      node={node}
-      valueType={valueType}
-    />
+      return !!getComponentDef(node.intermediateItem.type).getOptionsEffectValueType();
+    }),
   );
-}
-
-function RunVisibleOptionsEffects({ valueType, node }: RunOptionEffectsProps) {
-  const isReadOnly = FormStore.useIsReadOnly();
-  const item = node.intermediateItem as CompIntermediate<CompWithBehavior<'canHaveOptions'>>;
-  const lookups = FormStore.bootstrap.useLayoutLookups();
-  const dataModelBindings = item.dataModelBindings as IDataModelBindingsOptionsSimple | undefined;
-  const groupBindings = item.dataModelBindings as
-    | IDataModelBindingsForGroupCheckbox
-    | IDataModelBindingsForGroupMultiselect;
-  const { unsorted, isFetching, downstreamParameters } = useFetchOptions({ item });
-  const { options, preselectedOption } = useFilteredAndSortedOptions({ unsorted, valueType, item });
-
-  if (isFetching || isReadOnly) {
-    // No need to run effects while fetching or if the data has not been set yet
-    return false;
-  }
-
-  // Quickfix to fix simpleBinding being cleared as stale in FileUploadWithTag,
-  // we don't store option values here so it makes no sense to do this,
-  // consider solving this more elegantly in the future.
-  // AFAIK, stale values are not removed from attachment tags, maybe they should?
-  const parentComponent = node.parent.type === 'node' ? lookups.getComponent(node.parent.baseId) : undefined;
-  const shouldRemoveStaleValues =
-    parentComponent?.type !== 'FileUploadWithTag' && !('renderAsSummary' in item && item.renderAsSummary);
 
   return (
     <>
-      {shouldRemoveStaleValues && (
-        <EffectRemoveStaleValues
-          item={item}
-          parent={node.parent}
-          valueType={valueType}
-          options={options}
-        />
-      )}
-      {preselectedOption !== undefined && (
-        <EffectPreselectedOptionIndex
-          item={item}
-          parent={node.parent}
-          preselectedOption={preselectedOption}
-          valueType={valueType}
-          options={options}
-        />
-      )}
-      {downstreamParameters && dataModelBindings && dataModelBindings.metadata ? (
-        <EffectSetDownstreamParameters
-          item={item}
-          downstreamParameters={downstreamParameters}
-        />
-      ) : null}
-      {dataModelBindings && dataModelBindings.label && !!groupBindings.group ? (
-        <EffectStoreLabelInGroup
-          item={item}
-          parent={node.parent}
-          options={options}
-        />
-      ) : null}
-      {dataModelBindings && dataModelBindings.label && !groupBindings.group ? (
-        <EffectStoreLabel
-          item={item}
-          parent={node.parent}
-          valueType={valueType}
-          options={options}
-        />
-      ) : null}
+      {nodes.map((node) => (
+        <DataModelLocationProviderFromRowContexts
+          key={node.id}
+          rowContexts={node.rowContexts}
+        >
+          <RunOptionsEffectsForNode
+            node={node}
+            valueType={getComponentDef(node.intermediateItem.type).getOptionsEffectValueType()!}
+          />
+        </DataModelLocationProviderFromRowContexts>
+      ))}
     </>
   );
 }
